Fixed touchableopacity issue and separated audio player
